requirements.txt
telebot
schedule
python-dotenv
